Переносит логику создания графиков статистики в контроллер.
