+++ src/main.js
@@ -3,17 +3,19 @@ import Store from './api/store.js';
 import Provider from './api/provider.js';
 import TripController from './controllers/trip-controller.js';
 import FilterController from './controllers/filter-controller.js';
+import StatisticsController from './controllers/statistics-controller';
 import EventsModel from './models/events-model.js';
 import EventModel from './models/event-model.js';
 import LoadEvents from './components/load-events.js';
 import MenuComponent from './components/menu.js';
-import StatisticsComponent from './components/statistics.js';
 import {renderComponent, removeComponent, RenderPosition} from './utils/render.js';
 import {MenuItem, AUTHORIZATION, END_POINT, STORE_NAME} from './const.js';
 
 const tripControlsElement = document.querySelector(`.trip-main__trip-controls`);
 const tripEventsElement = document.querySelector(`.trip-events`);
 const pageMainElement = document.querySelector(`.page-main`);
+const bodyContainerElement = pageMainElement.querySelector(`.page-body__container`);
+const eventAddButtonElement = document.querySelector(`.trip-main__event-add-btn`);
 
 window.addEventListener(`load`, () => {
   navigator.serviceWorker.register(`/sw.js`)
@@ -40,31 +42,27 @@ const menuItems = Object.values(MenuItem)
 const menuComponent = new MenuComponent(menuItems);
 const filterController = new FilterController(tripControlsElement, eventsModel);
 const tripController = new TripController(tripEventsElement, eventsModel, apiWithProvider);
+const statisticsController = new StatisticsController(bodyContainerElement, eventsModel);
 const loadEvents = new LoadEvents();
-const statisticsComponent = new StatisticsComponent(eventsModel);
 
 renderComponent(tripControlsElement, menuComponent, RenderPosition.AFTERBEGIN);
 filterController.render();
 renderComponent(tripEventsElement, loadEvents);
-renderComponent(pageMainElement.querySelector(`.page-body__container`), statisticsComponent);
 
-statisticsComponent.hide();
-
-document.querySelector(`.trip-main__event-add-btn`)
-  .addEventListener(`click`, () => {
-    tripController.createEvent();
-  });
+eventAddButtonElement.addEventListener(`click`, () => {
+  tripController.createEvent();
+});
 
 menuComponent.setItemClickHandler((menuItem) => {
   switch (menuItem) {
     case MenuItem.STATS:
       menuComponent.setActiveItem(MenuItem.STATS);
       tripController.hide();
-      statisticsComponent.show();
+      statisticsController.show();
       break;
     case MenuItem.TABLE:
       menuComponent.setActiveItem(MenuItem.TABLE);
-      statisticsComponent.hide();
+      statisticsController.hide();
       tripController.show();
       break;
   }
@@ -83,6 +81,8 @@ Promise.all([
     eventsModel.setEvents(events);
     tripController.render();
     removeComponent(loadEvents);
+    statisticsController.render();
+    statisticsController.hide();
   });
 
 window.addEventListener(`online`, () => {
